Add files via upload
This is version 4.0 of the file transfer program. The GUI  has been considerably improved, the user now initiates the transfers, and the program only transfers new files created since the previous transfer. A copy of the LastCopy database, which holds the timestamp for all the previous transfers, has been included.