Add files via upload
This is a simple dart game simulation, which has no bells and whistles, not even graphics. It scores the event until a winner reaches 300, then awards the win to the highest score. Note that it has had server problems (2 of the 3 servers it was tried on failed on the first pass between libraries, the third failed to reinitialize the random correctly, but an identical program by Bob Tabor ran perfectly on recompile. I suggest a recompile using Visual Studio before attempting this program.