Add files via upload
This is a comic book inventory  program. It uses Visual Studio MVC architecture. The original version of this was by Bob Tabor of Dev U, but my version has additional functionality, new titles (Marvel superheroes were not in the data before) and a fictional company for copyrighting the software. It needs no outside database for the demonstration, but could be modified to use either client or server-side databases.