Add files via upload
This is a prototype for an app I'm developing. It can give basic information on how your dairy operation is going when handling a small herd of dairy cattle. Due to a problem with the USDA pricing database, some data has been emulated. Future versions will be fully interactive with that data source.